changed databases to actaully be a provider like it should be made guildtier multiplier be doubles changed guildhandler tier object loading to now use tiers.list rather than just tiers addedsome todos to certain classes with notes on how this is wrong/should be changed random edits

+++ src/main/java/me/glaremasters/guilds/guild/GuildTier.java
@@ -15,8 +15,8 @@ public class GuildTier {
     private final String name;
     private final double cost;
     private final int maxMembers;
-    private final int mobXpMultiplier;
-    private final int damageMultiplier;
+    private final double mobXpMultiplier;
+    private final double damageMultiplier;
     private final double maxBankBalance;
     private final int membersToRankup;
     private final List<String> permissions;
